Extract MockAction class to separate file for better modularity
