+++ src/sherpa_ai/actions/__init__.py
@@ -2,6 +2,7 @@ from sherpa_ai.actions.arxiv_search import ArxivSearch
 from sherpa_ai.actions.deliberation import Deliberation
 from sherpa_ai.actions.empty import EmptyAction
 from sherpa_ai.actions.google_search import GoogleSearch
+from sherpa_ai.actions.mockaction import MockAction
 from sherpa_ai.actions.planning import TaskPlanning
 from sherpa_ai.actions.synthesize import SynthesizeOutput
 
@@ -11,5 +12,6 @@ __all__ = [
     "TaskPlanning",
     "SynthesizeOutput",
     "ArxivSearch",
-    "EmptyAction"
+    "EmptyAction",
+    "MockAction"
 ]
